move chainbase related methods from framework to chainbase module.
